Refactor betty.project.new into an API
